+++ src/Believe/Core/ModelBase.cs
@@ -3,6 +3,7 @@ using Believe.Exceptions;
 using Believe.Models.Believe;
 using Believe.Models.Biscuits;
 using Believe.Models.Characters;
+using Believe.Models.Client.TicketSales;
 using Believe.Models.Conflicts;
 using Believe.Models.Matches;
 using Believe.Models.Quotes;
@@ -72,6 +73,13 @@ public abstract record class ModelBase
             new ApiEnumConverter<string, Webhooks::MemberType>(),
             new ApiEnumConverter<string, Webhooks::TransferType>(),
             new ApiEnumConverter<string, Webhooks::TeamMemberTransferredEventType>(),
+            new ApiEnumConverter<string, TicketSaleCreateResponsePurchaseMethod>(),
+            new ApiEnumConverter<string, TicketSaleRetrieveResponsePurchaseMethod>(),
+            new ApiEnumConverter<string, TicketSaleUpdateResponsePurchaseMethod>(),
+            new ApiEnumConverter<string, TicketSaleListResponsePurchaseMethod>(),
+            new ApiEnumConverter<string, PurchaseMethod>(),
+            new ApiEnumConverter<string, TicketSaleUpdateParamsPurchaseMethod>(),
+            new ApiEnumConverter<string, TicketSaleListParamsPurchaseMethod>(),
         },
     };