+++ src/Believe/Services/IConflictService.cs
@@ -7,9 +7,11 @@ using Believe.Models.Conflicts;
 namespace Believe.Services;
 
 /// <summary>
-/// NOTE: Do not inherit from this type outside the SDK unless you're okay with breaking
-/// changes in non-major versions. We may add new methods in the future that cause
-/// existing derived classes to break.
+/// Interactive endpoints for motivation and guidance
+///
+/// <para>NOTE: Do not inherit from this type outside the SDK unless you're okay with
+/// breaking changes in non-major versions. We may add new methods in the future that
+/// cause existing derived classes to break.</para>
 /// </summary>
 public interface IConflictService
 {