+++ src/Believe/Services/IPepTalkService.cs
@@ -7,9 +7,11 @@ using Believe.Models.PepTalk;
 namespace Believe.Services;
 
 /// <summary>
-/// NOTE: Do not inherit from this type outside the SDK unless you're okay with breaking
-/// changes in non-major versions. We may add new methods in the future that cause
-/// existing derived classes to break.
+/// Server-Sent Events (SSE) streaming endpoints
+///
+/// <para>NOTE: Do not inherit from this type outside the SDK unless you're okay with
+/// breaking changes in non-major versions. We may add new methods in the future that
+/// cause existing derived classes to break.</para>
 /// </summary>
 public interface IPepTalkService
 {