+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 0.2.0 (2026-03-09)
+
+Full Changelog: [v0.1.0...v0.2.0](https://github.com/cjavdev/believe-csharp/compare/v0.1.0...v0.2.0)
+
+### Features
+
+* **api:** add ticket sales endpoints, update biscuit/quote types ([b039343](https://github.com/cjavdev/believe-csharp/commit/b03934370492b4489a9c47cc25061cf4f62d6fb0))
+* **api:** manual updates ([78367fd](https://github.com/cjavdev/believe-csharp/commit/78367fd537ec91b25afee3dce0fa5a50063d4623))
+* **api:** manual updates ([fef552a](https://github.com/cjavdev/believe-csharp/commit/fef552af8c08cc5771cae9881e037985e07490e2))
+
+
+### Chores
+
+* **docs:** add proxy documentation to readme ([78ad97b](https://github.com/cjavdev/believe-csharp/commit/78ad97ba251eecb219423368b8e7a09e006b2ae5))
+* **docs:** add undocumented parameters to readme ([bb8bd9c](https://github.com/cjavdev/believe-csharp/commit/bb8bd9c5911d4d46b03f7b4d72f12d3ca469f70e))
+* **internal:** codegen related update ([ee68897](https://github.com/cjavdev/believe-csharp/commit/ee6889754ddfec16be6ad9c64457f2313f7b071b))
+* **tests:** add tests for retry logic ([807a352](https://github.com/cjavdev/believe-csharp/commit/807a35234eea881d73aa7fb0ef52a77e75f47ff7))
+* update placeholder string ([4b3b9b0](https://github.com/cjavdev/believe-csharp/commit/4b3b9b0c0cffea866dc54906cf22594c975c9fc4))
+
+
+### Documentation
+
+* remove typo in README.md ([58e1056](https://github.com/cjavdev/believe-csharp/commit/58e1056137b720be856a0933797e2907b494bf85))
+
 ## 0.1.0 (2026-02-21)
 
 Full Changelog: [v0.0.1...v0.1.0](https://github.com/cjavdev/believe-csharp/compare/v0.0.1...v0.1.0)